Improve command line friendliness (#104)
* Add support for `--key=value` command line arguments

This adds support for the `--key=value` syntax, in addition to the existing`--key value` and `--key` syntaxes, for a slightly more user-friendly experience when setting the command line.

* Add a command line help argument

Prints information about all of the commandline arguments and their options.  A lot of the information is transcribed from the github docs, and hopefully is factual.

* Formatting changes to command line help

* Update command line help message.

We now only print a very basic command line argument list, and tell the user to reference the online documentation for additional information.